fix : spotless, checkStyle 에러 해결 및 디렉토리 구조 변경
