+++ src/main/java/hotspot/user/UserApplication.java
@@ -6,8 +6,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class UserApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(UserApplication.class, args);
-	}
-
+  public static void main(String[] args) {
+    SpringApplication.run(UserApplication.class, args);
+  }
 }
