--- src/main/java/hotspot/user/common/ex/controller/port/FindExService.java
@@ -1,8 +1,0 @@
-package hotspot.user.common.ex.controller.port;
-
-import hotspot.user.common.ex.controller.response.ExResponse;
-
-public interface FindExService {
-
-    ExResponse findEx(Long exId);
-}
