--- src/main/java/hotspot/user/common/ex/controller/port/SaveExService.java
@@ -1,8 +1,0 @@
-package hotspot.user.common.ex.controller.port;
-
-import hotspot.user.common.ex.controller.request.CreateExRequest;
-
-public interface SaveExService {
-
-    void save(CreateExRequest request);
-}
