--- src/main/java/hotspot/user/common/ex/controller/port/UpdateExService.java
@@ -1,8 +1,0 @@
-package hotspot.user.common.ex.controller.port;
-
-import hotspot.user.common.ex.controller.request.UpdateExRequest;
-
-public interface UpdateExService {
-
-    void update(long exId, UpdateExRequest request);
-}
