--- src/main/java/hotspot/user/common/ex/controller/request/CreateExRequest.java
@@ -1,7 +1,0 @@
-package hotspot.user.common.ex.controller.request;
-
-public record CreateExRequest(
-        String exName,
-        String exDescription
-) {
-}
