--- src/main/java/hotspot/user/common/ex/controller/request/UpdateExRequest.java
@@ -1,7 +1,0 @@
-package hotspot.user.common.ex.controller.request;
-
-public record UpdateExRequest(
-        String exName,
-        String exDescription
-) {
-}
