--- src/main/java/hotspot/user/common/ex/controller/response/ExResponse.java
@@ -1,8 +1,0 @@
-package hotspot.user.common.ex.controller.response;
-
-public record ExResponse(
-        long exId,
-        String exName,
-        String exDescription
-) {
-}
