--- src/main/java/hotspot/user/common/ex/infrastructure/ExJpaRepository.java
@@ -1,7 +1,0 @@
-package hotspot.user.common.ex.infrastructure;
-
-import hotspot.user.common.ex.infrastructure.entity.ExEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ExJpaRepository extends JpaRepository<ExEntity, Long> {
-}
