+++ src/main/java/hotspot/user/common/exception/ApplicationException.java
@@ -4,7 +4,7 @@ import hotspot.user.common.exception.code.BaseErrorCode;
 
 public class ApplicationException extends BaseException {
 
-    public ApplicationException(BaseErrorCode code) {
-        super(code);
-    }
+  public ApplicationException(BaseErrorCode code) {
+    super(code);
+  }
 }
