+++ src/main/java/hotspot/user/common/exception/BaseException.java
@@ -6,18 +6,18 @@ import lombok.Getter;
 @Getter
 public abstract class BaseException extends RuntimeException {
 
-    private final BaseErrorCode code;
+  private final BaseErrorCode code;
 
-    protected BaseException(BaseErrorCode code) {
-        super(code.getMessage());
-        this.code = code;
-    }
+  protected BaseException(BaseErrorCode code) {
+    super(code.getMessage());
+    this.code = code;
+  }
 
-    public static <T extends BaseException> T from(BaseErrorCode code, Class<T> exceptionClass) {
-        try {
-            return exceptionClass.getConstructor(BaseErrorCode.class).newInstance(code);
-        } catch (Exception e) {
-            throw new RuntimeException("Could not create exception instance", e);
-        }
+  public static <T extends BaseException> T from(BaseErrorCode code, Class<T> exceptionClass) {
+    try {
+      return exceptionClass.getConstructor(BaseErrorCode.class).newInstance(code);
+    } catch (Exception e) {
+      throw new RuntimeException("Could not create exception instance", e);
     }
+  }
 }
