+++ src/main/java/hotspot/user/ex/controller/port/FindExService.java
@@ -1,0 +1,8 @@
+package hotspot.user.ex.controller.port;
+
+import hotspot.user.ex.controller.response.ExResponse;
+
+public interface FindExService {
+
+  ExResponse findEx(Long exId);
+}
