+++ src/main/java/hotspot/user/ex/controller/port/SaveExService.java
@@ -1,0 +1,8 @@
+package hotspot.user.ex.controller.port;
+
+import hotspot.user.ex.controller.request.CreateExRequest;
+
+public interface SaveExService {
+
+  void save(CreateExRequest request);
+}
