+++ src/main/java/hotspot/user/ex/controller/port/UpdateExService.java
@@ -1,0 +1,8 @@
+package hotspot.user.ex.controller.port;
+
+import hotspot.user.ex.controller.request.UpdateExRequest;
+
+public interface UpdateExService {
+
+  void update(long exId, UpdateExRequest request);
+}
