+++ src/main/java/hotspot/user/ex/controller/request/CreateExRequest.java
@@ -1,0 +1,3 @@
+package hotspot.user.ex.controller.request;
+
+public record CreateExRequest(String exName, String exDescription) {}
