+++ src/main/java/hotspot/user/ex/controller/request/UpdateExRequest.java
@@ -1,0 +1,3 @@
+package hotspot.user.ex.controller.request;
+
+public record UpdateExRequest(String exName, String exDescription) {}
