+++ src/main/java/hotspot/user/ex/controller/response/ExResponse.java
@@ -1,0 +1,3 @@
+package hotspot.user.ex.controller.response;
+
+public record ExResponse(long exId, String exName, String exDescription) {}
